Add eslint-patch and move dependencies
As we have shared the ESLint config, we want plugins to be included,
otherwise we have to install them manually in every project.

@rushstack/eslint-patch is used to include plugins dependencies instead
of imposing them as peer dependencies.

The creator of ESLint came up with the solution after several years,
which is experimental, so let's stick to the proven solution.
https://github.com/eslint/eslint/issues/3458#issuecomment-1239811334